+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Use monotonic high resolution timer for worker tasks run time.
 - Add worker tasks median run time to statistics.
-- Add worker tasks queue.
+- Add worker tasks queue (experimental).
 
 ## [2.4.4] - 2023-04-07